refactored event map and some more sailboat refectoring
