+++ Sailboat/Sources/Page/Helpers/EventResult.swift
@@ -5,6 +5,7 @@
 //  Created by Joshua Davis on 12/17/23.
 //
 
+// TODO: remove this alias its confusing
 public typealias Events = [String: (EventResult) -> Void]
 
 public enum EventResult {
@@ -12,5 +13,13 @@ public enum EventResult {
          float(Double),
          bool(Bool),
          int(Int),
-         none
+         none // TODO: rename to empty to not conflict with optional
 }
+
+//public enum EventResultTypes {
+//    case string,
+//         float,
+//         bool,
+//         int,
+//         empty
+//}
