+++ Sailboat/Sources/PageNode/PageNodes/CustomNode.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  CustomNode.swift
 //
 //
 //  Created by Joshua Davis on 12/22/23.
