+++ Sailboat/Sources/PageNode/PageNodes/OperatorNode.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  OperatorNode.swift
 //  
 //
 //  Created by Joshua Davis on 12/22/23.
