--- Sailor/Sources/SailboatSupport/Page+className.swift
@@ -1,20 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Joshua Davis on 10/6/23.
-//
-
-
-// TODO: depricate this?
-//extension Page {
-//    // TODO: these can be auto generated
-//    public func className(_ value: String) -> Self {
-//        return self.attribute(.className, value: value)
-//    }
-//
-//    public func src(_ value: String) -> Self {
-//        return self.attribute(.src, value: value)
-//
-//    }
-//}
