+++ Sailor/Sources/SailboatSupport/SailboatTypes.swift
@@ -7,15 +7,6 @@
 
 import Sailboat
 
-//public typealias Style = Sailboat.Style
-
 public typealias Page = Sailboat.Page
 public typealias State = Sailboat.State
 public typealias Binding = Sailboat.Binding
-
-//public typealias Unit = Sailboat.Unit
-
-//public typealias Div = Sailboat.Div
-//public typealias Span = Sailboat.Span
-//public typealias Button = Sailboat.Button
-// TODO: More
